convert src directory to typescript
including adding type annotations for everything as well as some
refactoring/api changes to make it fit better with static types.

why typescript? i really like typescript and wanted to al least
publish a typescript definition file with the package. it will
be much easier to keep in sync if i write the src in typescript
as well. it works out really nice for a lot of things. for
example: directional channels (similar to golang) can now be
acheived and checked at compile time by using specialized
directional interface types `InChannel<T>` and `OutChannel<T>`.

chan will continue to be just as usable from javascript.

updates to the build and tests as well as removal of the old
babel dependencies will come in future commits. this is a big
enough chunck already.